Re-locate all EVMEvents related files to `trnsp-custom` package
