+++ packages/apps/public/locales/en/app-explorer.json
@@ -2,7 +2,6 @@
   "API stats": "API stats",
   "Block details": "Block details",
   "Chain info": "Chain info",
-  "EVM info": "EVM info",
   "Forks": "Forks",
   "Latency": "Latency",
   "No blocks available": "No blocks available",
