+++ packages/page-explorer/src/index.tsx
@@ -5,17 +5,16 @@ import type { TFunction } from 'i18next';
 import type { TabItem } from '@polkadot/react-components/Tabs/types';
 import type { KeyedEvent } from '@polkadot/react-hooks/ctx/types';
 
+import { useBlockEVMEvents } from '@trnsp/custom/hooks/useBlockEVMEvents';
 import React, { useMemo, useRef } from 'react';
 import { Route, Switch } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 import { useApi, useBlockAuthors, useBlockEvents } from '@polkadot/react-hooks';
-// import { useBlockEVMEvents } from '@polkadot/react-hooks/useBlockEVMEvents';
 import { isFunction } from '@polkadot/util';
 
 import Api from './Api';
 import BlockInfo from './BlockInfo';
-import EVM from './EVM';
 import Forks from './Forks';
 import Latency from './Latency';
 import Main from './Main';
@@ -50,10 +49,6 @@ function createItemsRef (t: TFunction): TabItem[] {
       text: t<string>('Chain info')
     },
     {
-      name: 'evm',
-      text: t<string>('EVM info')
-    },
-    {
       hasParams: true,
       name: 'query',
       text: t<string>('Block details')
@@ -83,6 +78,7 @@ function ExplorerApp ({ basePath, className }: Props): React.ReactElement<Props>
   const { api } = useApi();
   const { lastHeaders } = useBlockAuthors();
   const { eventCount, events } = useBlockEvents();
+  const evmEvents = useBlockEVMEvents();
 
   const itemsRef = useRef(createItemsRef(t));
   const pathRef = useRef(createPathRef(basePath));
@@ -100,7 +96,6 @@ function ExplorerApp ({ basePath, className }: Props): React.ReactElement<Props>
         items={itemsRef.current}
       />
       <Switch>
-        <Route path={pathRef.current.evm}><EVM /></Route>
         <Route path={pathRef.current.api}><Api /></Route>
         <Route path={pathRef.current.forks}><Forks /></Route>
         <Route path={pathRef.current.latency}><Latency /></Route>
@@ -110,6 +105,7 @@ function ExplorerApp ({ basePath, className }: Props): React.ReactElement<Props>
           <Main
             eventCount={eventCount}
             events={events}
+            evmEvents={evmEvents}
             headers={lastHeaders}
           />
         </Route>
