+++ packages/react-components/src/index.tsx
@@ -38,7 +38,6 @@ export { default as Event } from './Event';
 export { default as Expander } from './Expander';
 export { default as ExpanderScroll } from './ExpanderScroll';
 export { default as Extrinsic } from './Extrinsic';
-export { default as EVMEventDisplay } from './EVM';
 export { default as FilterOverlay } from './FilterOverlay';
 export { default as Flag } from './Flag';
 export { default as Forget } from './Forget';
