+++ packages/trnsp-custom/src/components/EVMEvents.tsx
@@ -2,17 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React, { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
 
-import BlockHeaders from '@polkadot/app-explorer/BlockHeaders';
-import Query from '@polkadot/app-explorer/Query';
-import Summary from '@polkadot/app-explorer/Summary';
-import { Columar, MarkError, Table } from '@polkadot/react-components';
-import { useBlockAuthors } from '@polkadot/react-hooks';
-import { BlockEVMEvents } from '@polkadot/react-hooks/ctx/types';
+import { MarkError, Table } from '@polkadot/react-components';
 
+import { BlockEVMEvents } from '../types';
 import EVMEvent from './EVMEvent';
-import { useTranslation } from './translate';
 
 interface Props {
   className?: string;
@@ -52,38 +48,25 @@ function EVMEvents ({ className = '', emptyLabel, error, eventClassName, events,
     ],
     [label, t]
   );
-  const { lastHeaders } = useBlockAuthors();
 
   return (
-    <>
-      <Query />
-      <Summary eventCount={0} />
-      <Columar>
-        <Columar.Column>
-          <BlockHeaders headers={lastHeaders} />
-        </Columar.Column>
-        <Columar.Column>
-
-          <Table
-            className={className}
-            empty={emptyLabel || t<string>('No events available')}
-            header={header}
+    <Table
+      className={className}
+      empty={emptyLabel || t<string>('No events available')}
+      header={header}
+    >
+      {error
+        ? (
+          <tr
+            className={eventClassName}
+            key='error'
           >
-            {error
-              ? (
-                <tr
-                  className={eventClassName}
-                  key='error'
-                >
-                  <td><MarkError content={t<string>('Unable to decode the block events. {{error}}', { replace: { error: error.message } })} /></td>
-                </tr>
-              )
-              : events && events.map((e) => renderEvent(eventClassName, e))
-            }
-          </Table>
-        </Columar.Column>
-      </Columar>
-    </>
+            <td><MarkError content={t<string>('Unable to decode the block events. {{error}}', { replace: { error: error.message } })} /></td>
+          </tr>
+        )
+        : events && events.map((e) => renderEvent(eventClassName, e))
+      }
+    </Table>
   );
 }
