+++ packages/trnsp-custom/src/hooks/useBlockEVMEvents.ts
@@ -1,14 +1,12 @@
 // Copyright 2017-2023 @polkadot/react-hooks authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-// import type { BlockEvents } from './ctx/types';
-
 import { useContext } from 'react';
 
-import { BlockEVMEvents } from '@polkadot/react-hooks/ctx/types';
+import { createNamedHook } from '@polkadot/react-hooks';
 
-import { BlockEVMEventsCtx } from './ctx/BlockEVMEvents';
-import { createNamedHook } from './createNamedHook';
+import { BlockEVMEventsCtx } from '../providers/BlockEVMEvents';
+import { BlockEVMEvents } from '../types';
 
 function useBlockEVMEventsImpl (): BlockEVMEvents[] {
   return useContext(BlockEVMEventsCtx);
